+++ ugbase/lib_grid/algorithms/geom_obj_util/volume_util_impl.hpp
@@ -97,11 +97,11 @@ CalculateCenter(const VolumeVertices* vol, TVertexPositionAttachmentAccessor& aa
 //	init v with 0.
 	VecSet(v, 0);
 
-	uint numVrts = vol->num_vertices();
+	size_t numVrts = vol->num_vertices();
 	VolumeVertices::ConstVertexArray vrts = vol->vertices();
 
 //	sum up
-	for(uint i = 0; i < numVrts; ++i)
+	for(size_t i = 0; i < numVrts; ++i)
 	{
 		VecAdd(v, v, aaPosVRT[vrts[i]]);
 	}
@@ -125,12 +125,12 @@ CalculateCenter(const VolumeVertices* vol, TAAPosVRT& aaPos, TAAWeightVRT& aaWei
 //	init v with 0.
 	VecSet(v, 0);
 
-	uint numVrts = vol->num_vertices();
+	size_t numVrts = vol->num_vertices();
 	VolumeVertices::ConstVertexArray vrts = vol->vertices();
 
 //	sum up
 	weight_t totalWeight = 0;
-	for(uint i = 0; i < numVrts; ++i)
+	for(size_t i = 0; i < numVrts; ++i)
 	{
 		weight_t w = aaWeight[vrts[i]];
 		VecScaleAppend(v, w, aaPos[vrts[i]]);
